Refactor: MainView -> MainViewController로 통합
