+++ HarryPotterSeriesApp/Controller/MainViewController.swift
@@ -2,17 +2,40 @@
 import UIKit
 import SnapKit
 
-class MainViewController: UIViewController {
+class MainViewController: UIViewController, MainHeaderViewDelegate {
     
+    var bookData = [Book]()
+    var selectedSeries = 0
+    
+    let headerView = MainHeaderView()
+    let contentView = MainContentView()
     let dataService = DataService()
-    let mainView = MainView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.addSubview(mainView)
-        mainView.snp.makeConstraints { $0.edges.equalToSuperview() }
-        
+        setupUI()
+        setupConstraints()
         loadBooks()
+    }
+    
+    private func setupUI() {
+        view.backgroundColor = .systemBackground
+        
+        headerView.delegate = self
+        
+        view.addSubview(headerView)
+        view.addSubview(contentView)
+    }
+    
+    private func setupConstraints() {
+        headerView.snp.makeConstraints {
+            $0.top.leading.trailing.equalTo(view.safeAreaLayoutGuide).inset(20)
+        }
+        
+        contentView.snp.makeConstraints {
+            $0.top.equalTo(headerView.snp.bottom).offset(10)
+            $0.leading.trailing.bottom.equalTo(view.safeAreaLayoutGuide)
+        }
     }
 
     func loadBooks() {
@@ -23,7 +46,8 @@ class MainViewController: UIViewController {
                 switch result {
                 case .success(let books):
                     guard !books.isEmpty else { return }
-                    self.mainView.configure(books: books)
+                    self.bookData = books
+                    self.updateContent()
                     
                 case .failure(let error):
                     self.handleError(error)
@@ -32,10 +56,20 @@ class MainViewController: UIViewController {
         }
     }
     
+    private func updateContent() {
+        headerView.configure(books: bookData, selectedSeries: selectedSeries)
+        contentView.configure(book: bookData[selectedSeries], selectedSeries: selectedSeries)
+    }
+    
     func handleError(_ error: Error) {
         let alert = UIAlertController(title: "데이터 로딩 실패", message: error.localizedDescription, preferredStyle: .alert)
         alert.addAction(.init(title: "확인", style: .default))
         present(alert, animated: true)
+    }
+    
+    func didTapSeriesButton(seriesIndex: Int) {
+        selectedSeries = seriesIndex
+        updateContent()
     }
 }
